[TLSCLIENT] Split the TC_TLSCLIENT.py file into separate test files
